requirement.txt
django~=1.11.0
numpy
numpy-stl
pillow
